requirements.txt
connexion[swagger-ui] >= 2.6.0; python_version>="3.6"
# 2.3 is the last version that supports python 3.4-3.5
connexion[swagger-ui] <= 2.3.0; python_version=="3.5" or python_version=="3.4"
# connexion requires werkzeug but connexion < 2.4.0 does not install werkzeug
# we must peg werkzeug versions below to fix connexion
# https://github.com/zalando/connexion/pull/1044
werkzeug >= 0.16.1
swagger-ui-bundle >= 0.0.2
python_dateutil >= 2.6.0
setuptools >= 21.0.0
Flask==2.0.1
boto3==1.24.20
botocore==1.27.20
amazon-braket-default-simulator==1.6.2
# amazon-braket-ocean-plugin==1.0.10
amazon-braket-schemas==1.10.0
amazon-braket-sdk==1.25.2
qiskit==0.36.2
qiskit-aer==0.10.4
qiskit-ibmq-provider==0.19.1
qiskit-ignis==0.7.1
qiskit-terra==0.20.2
pyyaml>=6.0
requests>=2.28.2
